# specs/003-add-a-matching/contracts/api.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  version: 1.0.0
  description: |
    API contracts for the admin dashboard content management system.
    Extends existing blog and portfolio APIs with admin-specific operations.
  contact:
    name: Rohan Batra
    email: contact@rohanbatra.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://rohanbatra.com/api
    description: Production server
security:
  - ClerkAuth: []
paths:
...
  /admin/settings:
    get:
      summary: Get all site settings organized by category
      tags: [Admin Settings Management]
      security:
        - ClerkAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: isPublic
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminSiteSetting'
                  categories:
                    type: array
                    items:
                      type: string

    post:
      summary: Create a new site setting
      tags: [Admin Settings Management]
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSettingRequest'
      responses:
        '201':
          description: Setting created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSiteSetting'
  /admin/settings/{key}:
    put:
      summary: Update a site setting
      tags: [Admin Settings Management]
      security:
        - ClerkAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingRequest'
      responses:
        '200':
          description: Setting updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSiteSetting'

    delete:
      summary: Delete a site setting
      tags: [Admin Settings Management]
      security:
        - ClerkAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Setting deleted successfully
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      description: Clerk JWT token in Authorization header
  schemas:
...
    AdminSiteSetting:
      type: object
      properties:
        _id:
          type: string
        key:
          type: string
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
            - type: array
        type:
          type: string
          enum: [string, number, boolean, object, array]
        category:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
        validation:
          type: object
          properties:
            required:
              type: boolean
            minLength:
              type: integer
            maxLength:
              type: integer
            pattern:
              type: string
            options:
              type: array
              items:
                type: string
        updatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSettingRequest:
      type: object
      properties:
        key:
          type: string
          pattern: ^[a-z0-9-]+$
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
            - type: array
        type:
          type: string
          enum: [string, number, boolean, object, array]
        category:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
          default: false
        validation:
          type: object
      required: [key, value, type, category]
    UpdateSettingRequest:
      type: object
      properties:
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
            - type: array
        description:
          type: string
        isPublic:
          type: boolean
        validation:
          type: object
      required: [value]

    # Utility Schemas
...